pyproject.toml
[project]
name = "hang2la-skills"
version = "0.1.0"
description = "AI-driven video generation pipeline using agent skills"
requires-python = ">=3.11"
dependencies = [
    "moviepy>=2.0.0",
    "edge-tts>=7.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff]
line-length = 100
